Add functionality to store book objects in array
When the user inputs the information for a book, that book info
needs to be stored in an array that represents the library. The
array has been created, the constructor for the book object has
been implemented, and the function that stores the user's book
input into the array has been added.