Create function to display books on page
The books need to be displayed on the webpage as the user adds them
to the library. The displayBooks function now does this. It loops
over the books in the myLibrary array and takes the info from each
book, adds it to a card div, and then appends that div to the library
container element that was created in the index file. Some dummy books
were created to demonstrate this functionality.