Create dialog box to add new book information
When the New Book button is clicked, a dialog box needs to pop up
to allow the user to input book information. This dialog box has
been created. When the Add Book button is clicked, the form vlaidates
that all the fields were input and valid. The user can also click
Cancel and cancel the action of adding a book. For now, the book does
not yet show up in the library; this functionality still needs to be
added.