Add New Book display functionality
Before, the dialog box that allowed the user to input a new book's
information popped up, but when the user submitted the form, the
book was not added to the library display. Now, the new book appears
on the display. There is form validation to prevent the user from
inputting incomplete book entries. There is also a cancel button to
cancel the action and add no new books. Whether a new book is added
or the form is cancelled, the form is reset when it is closed.