Add functionality to delete book from library
It would be useful to the user to be able to delete a book from
their library. A Remove button has been added to each book card
as it is added to the library. When the button is clicked, the
index of that book is retrieved from the array, the book is removed
from the array, and that child node is removed from the library
container on the display.