Add method to Book to toggle read status
The user will be able to click a button to toggle the reading
status of each book in their library. In order to do this, there
needs to be a method in the Book object that toggles the book's
read status. This has been implemented and will be utilized in
future functionality.