Add button to each book to toggle read status
The user needs to be able to change the read status of each book
in their library as they read books. A button has been added to
each book's display that toggles the read status of that book and
displays the read status in its text content. The background color
also changes to differentiate the two statuses.