Add styling to header of page
Before, the webpage was composed of basic HTML elements with only
enough styling to distinguish elements from one another. Now, the
styling has been updated. The entire page's structure is represented
by a grid with a "navigation" header and a library section. The header
has been given some basic styling, including logos, a background color,
and improved spacing and font size. The header itself is a flex
container to make the placement of items easier as the webpage is
resized.